refactor: extract backend route cores
